Remove obsolete LICENSE file warning (code & tmpl)
8ffce20c0210a9cb7bfbbce99b5910c1cd8ea199 introduced a warning for
repositories that did not contain a license file and this filename check
was added.
3344064434921895721d2ae444731ad81eaa9b82 removed the warning, but the
filename check was not reverted.  Instead, the boolean
c.Data["LicenseExists"], which triggers the warning when false, was
fixed to true and the template was left as is.

The repo/view_list template code has been reverted as well.