[routes/repo] Minimise upstream diff renderDirectory()
Split out code for parsing datacite.yml file and for rendering annexed
README files on the directory listing page of the repository view.

New function 'resolveAnnexedContent()' takes a file content buffer and
an io.Reader as arguments and if the buffer contains annex link text,
returns the first 1024 bytes of the referenced annexed content as well
as a buffered bufio.Reader, otherwise it returns the original data.  The
returned byte slice and bufio.Reader should be used to replace the
buffer and io.Reader sent in through the caller so that the rest of the
task (checking if it's text, loading the rest of the data, etc) can be
handled by the original upstream 'renderDirectory()' function.

The new 'resolveAnnexedContent()' is intended to be used in all other
places where a blob being read (either for rendering or for download)
should be replaced by the annexed content it potentially points to.
This might require some generalisation to avoid loading very large files
into memory when not necessary.