[routes/repo] Minimise upstream diff renderFile()
Split out code for rendering files in the repository view.

Uses the existing 'resolveAnnexedContent()' to replace the buffer and
reader as in 'renderDirectory()'.  'resolveAnnexedContent()' now
additionally saves the size of the file in c.Data["FileSize"] to
properly compare with MaxDisplayFileSize.

Notes:
- CAPTCHAs are disabled for now for all files.  We will revisit the
CAPTCHA rule and readd a general function showing the challenge and
redirecting to the data.
- All non-text files are rendered by default.  The MaxDisplayFileSize is
only effective for text file rendering (from upstream).