Scan through files for datacite.yml separately
The original loop breaks after finding a readme file, so if the readme
is found before the datacite, the loop breaks and the datacite is never
found.  We could modify it so that it keeps running until the end, but
I don't want to make bigger changes to the existing upstream code.  It
would also change the behaviour when there are multiple files that are
detected as README files (which is a separate issue).