Make indexing call in a goroutine
If the indexing service is misconfigured, or down, or just slow,
responses to commits can hang until they timeout.  This affects both
git pushes from the command line as well as web uploads or file changes.

Since the response to an indexing request is not relevant to the user,
we should simply spawn a goroutine to make the call and return
immediately.