More efficient annex content location resolution
See https://github.com/G-Node/libgin/issues/9
The NewAFile in libgin runs a search through the annex object store for
the file matching the key. Shelling out to git annex with the
'contentlocation' command is constant time since it determines the
location directly based on the key.
We could also use the hashmix functions from libgin for this.