New file: gin-features.md
Describes features that have been added to GIN compared to upstream
GOGS.  In the future it might also point to specific parts
of the code where the changes are implemented.

This should be consulted when writing automated tests for GIN features
and reviewing the state of features when pulling in changes from
upstream.

Work in progress.