[CI] Update github action
Replacing the outdated, unavailable golangci with the latest version.

Both linters and shell checks would require file updates from
upstream gogs which is beyond the scope of this PR. The introduced
changes only enable the workflows to pass but upstream changes
need to be introduced to properly address the issues at a later
point.

Also adding go 1.16.x to the test matrix.