RF: in list of files, link to tree (not commit) for submodules
While navigating files it makes more sense to jump to that specific
state of the tree instead of specific commit diff.  With such trivial
change we should be able to achive that.  In other similar uses of /commit/
I left them as is since there it does makes more sense to jump to diff.

Inspired by @aksoo